Uplift of #10650 (squashed) to beta
