+++ components/brave_wallet/browser/brave_wallet_utils.cc
@@ -16,13 +16,13 @@
 #include "base/logging.h"
 #include "base/strings/string_number_conversions.h"
 #include "base/strings/string_split.h"
-#include "base/strings/string_util.h"
 #include "base/strings/stringprintf.h"
 #include "base/time/time.h"
 #include "base/values.h"
 #include "brave/components/brave_wallet/browser/pref_names.h"
 #include "brave/components/brave_wallet/common/brave_wallet.mojom.h"
 #include "brave/components/brave_wallet/common/features.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
 #include "brave/components/brave_wallet/common/value_conversion_utils.h"
 #include "brave/third_party/ethash/src/include/ethash/keccak.h"
 #include "brave/vendor/bip39wally-core-native/include/wally_bip39.h"
@@ -232,19 +232,6 @@ GetAllKnownNetworksForTesting() {
   return result;
 }
 
-std::string ToHex(const std::string& data) {
-  if (data.empty()) {
-    return "0x0";
-  }
-  return "0x" + base::ToLowerASCII(base::HexEncode(data.data(), data.size()));
-}
-
-std::string ToHex(const std::vector<uint8_t>& data) {
-  if (data.empty())
-    return "0x0";
-  return "0x" + base::ToLowerASCII(base::HexEncode(data));
-}
-
 std::string KeccakHash(const std::string& input, bool to_hex) {
   std::vector<uint8_t> bytes(input.begin(), input.end());
   std::vector<uint8_t> result = KeccakHash(bytes);
@@ -260,81 +247,6 @@ std::vector<uint8_t> KeccakHash(const std::vector<uint8_t>& input) {
 std::string GetFunctionHash(const std::string& input) {
   std::string result = KeccakHash(input);
   return result.substr(0, std::min(static_cast<size_t>(10), result.length()));
-}
-
-// Pads a hex encoded parameter to 32-bytes
-// i.e. 64 hex characters.
-bool PadHexEncodedParameter(const std::string& hex_input, std::string* out) {
-  if (!out) {
-    return false;
-  }
-  if (!IsValidHexString(hex_input)) {
-    return false;
-  }
-  if (hex_input.length() >= 64 + 2) {
-    *out = hex_input;
-    return true;
-  }
-  std::string hex_substr = hex_input.substr(2);
-  size_t padding_len = 64 - hex_substr.length();
-  std::string padding(padding_len, '0');
-
-  *out = "0x" + padding + hex_substr;
-  return true;
-}
-
-// Determines if the passed in hex string is valid
-bool IsValidHexString(const std::string& hex_input) {
-  if (hex_input.length() < 3) {
-    return false;
-  }
-  if (!base::StartsWith(hex_input, "0x")) {
-    return false;
-  }
-  for (const auto& c : hex_input.substr(2)) {
-    if (!base::IsHexDigit(c)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-// Takes 2 inputs prefixed by 0x and combines them into an output with a single
-// 0x. For example 0x1 and 0x2 would return 0x12
-bool ConcatHexStrings(const std::string& hex_input1,
-                      const std::string& hex_input2,
-                      std::string* out) {
-  if (!out) {
-    return false;
-  }
-  if (!IsValidHexString(hex_input1) || !IsValidHexString(hex_input2)) {
-    return false;
-  }
-  *out = hex_input1 + hex_input2.substr(2);
-  return true;
-}
-
-bool ConcatHexStrings(const std::vector<std::string>& hex_inputs,
-                      std::string* out) {
-  if (!out) {
-    return false;
-  }
-  if (hex_inputs.empty()) {
-    return false;
-  }
-  if (!IsValidHexString(hex_inputs[0])) {
-    return false;
-  }
-
-  *out = hex_inputs[0];
-  for (size_t i = 1; i < hex_inputs.size(); i++) {
-    if (!IsValidHexString(hex_inputs[i])) {
-      return false;
-    }
-    *out += hex_inputs[i].substr(2);
-  }
-
-  return true;
 }
 
 // Takes a hex string as input and converts it to a uint256_t
