+++ components/brave_wallet/browser/eth_data_builder.cc
@@ -7,6 +7,7 @@
 
 #include "base/logging.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
 
 namespace brave_wallet {
