+++ components/brave_wallet/browser/eth_requests.cc
@@ -8,7 +8,7 @@
 #include <utility>
 
 #include "base/json/json_writer.h"
-#include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
 
 namespace {
