+++ components/brave_wallet/browser/rlp_encode_unittest.cc
@@ -7,8 +7,8 @@
 #include <string>
 #include <utility>
 
-#include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
 #include "brave/components/brave_wallet/browser/rlp_encode.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
 namespace {
