+++ components/brave_wallet/common/BUILD.gn
@@ -1,5 +1,7 @@
+import("//brave/build/config.gni")
 import("//brave/components/brave_wallet/common/buildflags/buildflags.gni")
 import("//mojo/public/tools/bindings/mojom.gni")
+import("//testing/test.gni")
 import("//tools/grit/preprocess_if_expr.gni")
 
 preprocess_folder = "preprocessed"
@@ -12,6 +14,8 @@ static_library("common") {
     "eth_request_helper.h",
     "features.cc",
     "features.h",
+    "hex_utils.cc",
+    "hex_utils.h",
     "value_conversion_utils.cc",
     "value_conversion_utils.h",
   ]
@@ -47,3 +51,22 @@ mojom("mojom") {
     "//url/mojom:url_mojom_gurl",
   ]
 }
+
+source_set("unit_tests") {
+  testonly = true
+  if (brave_wallet_enabled) {
+    sources = [
+      "eth_request_helper_unittest.cc",
+      "hex_utils_unittest.cc",
+      "value_conversion_utils_unittest.cc",
+    ]
+
+    deps = [
+      ":common",
+      ":mojom",
+      "//base/test:test_support",
+      "//testing/gtest",
+      "//url",
+    ]
+  }  # if (brave_wallet_enabled)
+}  # source_set("brave_wallet_common_unit_tests")
