--- components/brave_wallet/common/test/BUILD.gn
@@ -1,26 +1,0 @@
-# Copyright (c) 2021 The Brave Authors. All rights reserved.
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this file,
-# You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-import("//brave/build/config.gni")
-import("//brave/components/brave_wallet/common/buildflags/buildflags.gni")
-import("//testing/test.gni")
-
-source_set("brave_wallet_common_unit_tests") {
-  testonly = true
-  if (brave_wallet_enabled) {
-    sources = [
-      "//brave/components/brave_wallet/common/eth_request_helper_unittest.cc",
-      "//brave/components/brave_wallet/common/value_conversion_utils_unittest.cc",
-    ]
-
-    deps = [
-      "//base/test:test_support",
-      "//brave/components/brave_wallet/common",
-      "//brave/components/brave_wallet/common:mojom",
-      "//testing/gtest",
-      "//url",
-    ]
-  }  # if (brave_wallet_enabled)
-}  # source_set("brave_wallet_common_unit_tests")
