+++ test/BUILD.gn
@@ -197,8 +197,8 @@ test("brave_unit_tests") {
     "//brave/components/brave_sync:crypto",
     "//brave/components/brave_sync:network_time_helper",
     "//brave/components/brave_wallet/browser/test:brave_wallet_unit_tests",
+    "//brave/components/brave_wallet/common:unit_tests",
     "//brave/components/brave_wallet/common/buildflags",
-    "//brave/components/brave_wallet/common/test:brave_wallet_common_unit_tests",
     "//brave/components/brave_wallet/renderer/test:unit_tests",
     "//brave/components/child_process_monitor:unittests",
     "//brave/components/ipfs/buildflags",
